--- a/lib/globalize.js
+++ b/lib/globalize.js
@@ -11,6 +11,7 @@ var debug = require('debug')('strong-globalize');
 var fs = require('fs');
 var helper = require('./helper');
 var osLocale = require('os-locale');
+var MapCache = require('lodash/_MapCache');
 var md5 = require('md5');
 var pathUtil = require('path');
 var translate = require('./translate');
@@ -128,7 +129,19 @@ function formatMessage(path, variables, lang) {
   }
   function formatMsgInline(lang) {
     var g = global.STRONGLOOP_GLB.bundles[lang];
-    return g.messageFormatter(path)(variables || {});
+    var allFormatters = global.STRONGLOOP_GLB.formatters;
+    var langFormatters;
+    if (allFormatters.has(lang)) {
+      langFormatters = allFormatters.get(lang);
+    } else {
+      langFormatters = new MapCache();
+      allFormatters.set(lang, langFormatters);
+    }
+    if (langFormatters.has(path))
+      return langFormatters.get(path)(variables || {});
+    var format = g.messageFormatter(path);
+    langFormatters.set(path, format);
+    return format(variables || {});
   }
   try {
     message = formatMsgInline(lang);
@@ -368,6 +381,7 @@ function loadGlobalize(lang, enumerateNodeModules) {
       versionSG: versionSG,
       versionG: versionG,
       bundles: {},
+      formatters: new MapCache(),
       load: Globalize.load,
       locale: Globalize.locale,
       loadMessages: Globalize.loadMessages,