--- a/lib/globalize.js
+++ b/lib/globalize.js
@@ -13,6 +13,7 @@ var helper = require('./helper');
 var osLocale = require('os-locale');
 var MapCache = require('lodash/_MapCache');
 var md5 = require('md5');
+var memoize = require('lodash/memoize');
 var pathUtil = require('path');
 var translate = require('./translate');
 var util = require('util');
@@ -117,9 +118,10 @@ function setDefaultLanguage(lang) {
  */
 function formatMessage(path, variables, lang) {
   assert(path);
-  var message = path;
-  if (helper.hashKeys(path)) path = md5(path);
   if (!global.STRONGLOOP_GLB) setDefaultLanguage();
+  var message = path;
+  if (helper.hashKeys(path))
+    path = global.STRONGLOOP_GLB.getHash(path);
   lang = lang || global.STRONGLOOP_GLB.DEFAULT_LANG;
   debug('~~~ lang = %s %s %j %s', lang, path, variables, __filename);
   var trailer = helper.getTrailerAfterDot(path);
@@ -382,6 +384,7 @@ function loadGlobalize(lang, enumerateNodeModules) {
       versionG: versionG,
       bundles: {},
       formatters: new MapCache(),
+      getHash: memoize(md5),
       load: Globalize.load,
       locale: Globalize.locale,
       loadMessages: Globalize.loadMessages,