--- a/lib/translate.js
+++ b/lib/translate.js
@@ -85,6 +85,12 @@ function loadMsgFromFile(lang, rootDir, enumerateNodeModules) {
       messages[lang] = jsonObj;
       global.STRONGLOOP_GLB.loadMessages(messages);
       helper.registerResTag(fileIdHash, fileName, lang, tagType);
+      if (global.STRONGLOOP_GLB.formatters.has(lang)) {
+        var formatters = global.STRONGLOOP_GLB.formatters.get(lang);
+        for (var key in jsonObj) {
+          formatters.delete(key);
+        }
+      }
     });
 }