Over the weekend, I have realised that the initial version of [format benchmark](https://github.com/strongloop/strong-globalize/blob/0b553dbbab4f931d6561c29ca63fadd97ec7beb4/benchmark/format.js) was measuring a wrong scenario. Instead of measuring how long it takes to serve a translated string from `message/{lang}/messages.js`, we were measuring how long it takes to translate a string that's not included in the messages. The first commit fixes that by adding another run to the benchmark file.

The second commit improves the performance by re-using message formatters created by `globalize` module, which is the recommended approach per https://www.npmjs.com/package/globalize#performance

In the last commit, I am adding a cache of `md5` hashes to avoid potentially slow computation.

I believe all of these changes are fully backwards compatible and should not break any existing applications.

The only downside I am aware of is that these two caches are slightly increasing memory usage. If this becomes ever a problem, we can limit the cache size by using e.g. [lru-cache](https://www.npmjs.com/package/lru-cache).

Connect to #66 

@Setogit please review
cc @rmg @sam-github 

Benchmark results - before:

```
$ node benchmark/format.js
BASELINE
  1940 calls of "util.format()" took 3.179012 milliseconds
CASE 1 - no messages are loaded
  1940 calls of "g.f()" took 818.321937 milliseconds
  g.f() is 258x slower than util.format
CASE 2 - all messages are loaded
  1940 calls of "g.f()" took 2197.188276 milliseconds
  g.f() is 692x slower than util.format
```

After:

```
BASELINE
  1940 calls of "util.format()" took 3.591508ms
CASE 1 - no messages are loaded
  1940 calls of "g.f()" took 695.200518ms
  g.f() is 194x slower than util.format
CASE 2 - all messages are loaded
  1940 calls of "g.f()" took 1114.674444ms
  g.f() is 311x slower than util.format```
```

The exact ratio differs across runs, probably depending on CPU state, etc. The consistent part is that CASE 2 completes in about 1 second now, while it took over 2 seconds before.