@rmg 2016-08-19
@bajtos if I understand correctly, you're basically memoizing `md5()`? If so, couldn't this be accomplished in a much simpler manner, possibly with equal or better performance, by something like this?
```js
var _ = require('lodash');
var md5 = _.memoize(require('md5'));
```

I recently read through the `_.memoize` code and it tries to use native `Map` instead of `Object.create(null)` when possible. It is also possible to override the resolver and the cache constructor, which would allow us to use LRU or a `WeakMap` (which has the advantage of releasing memory only when there is enough VM pressure to GC the entries).