@bajtos 2016-08-23
> if I understand correctly, you're basically memoizing md5()? If so, couldn't this be accomplished in a much simpler manner...

@rmg good point, I have reworked both `md5` memoization and formatter's cache to use `memoize` and `MapCache` instead of `Object.create(null)`.

@Setogit could you please review?